# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "JustoAI V2 API",
    "version": "2.0.0",
    "description": "API completa da plataforma JustoAI V2 para gestão jurídica com IA",
    "contact": {
      "name": "Equipe JustoAI",
      "email": "dev@justoai.com"
    }
  },
  "servers": [
    {
      "url": "http://localhost:3000",
      "description": "Servidor de Desenvolvimento"
    }
  ],
...
  "paths": {
...
    "/api/reports/generate": {
      "post": {
        "summary": "Gerar relatório",
        "description": "Gera um novo relatório baseado nos parâmetros fornecidos",
        "tags": [
          "Relatórios"
        ],
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": [
                  "type",
                  "format"
                ],
                "properties": {
                  "type": {
                    "type": "string",
                    "enum": [
                      "PROCESS_SUMMARY",
                      "CLIENT_REPORT",
                      "CASE_ANALYSIS",
                      "FINANCIAL"
                    ],
                    "description": "Tipo de relatório"
                  },
                  "format": {
                    "type": "string",
                    "enum": [
                      "PDF",
                      "DOCX",
                      "XLSX"
                    ],
                    "description": "Formato do relatório"
                  },
                  "filters": {
                    "type": "object",
                    "properties": {
                      "dateFrom": {
                        "type": "string",
                        "format": "date"
                      },
                      "dateTo": {
                        "type": "string",
                        "format": "date"
                      },
                      "clientIds": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      },
                      "caseIds": {
                        "type": "array",
                        "items": {
                          "type": "string"
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Relatório gerado com sucesso",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "object",
                      "properties": {
                        "reportId": {
                          "type": "string",
                          "format": "uuid"
                        },
                        "status": {
                          "type": "string",
                          "example": "PROCESSING"
                        },
                        "downloadUrl": {
                          "type": "string",
                          "format": "uri",
                          "nullable": true
                        }
                      }
                    },
                    "message": {
                      "type": "string",
                      "example": "Relatório em processamento"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/ValidationError"
          },
          "401": {
            "$ref": "#/components/responses/UnauthorizedError"
          }
        }
      }
    },
...
  },
...
}